+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Add liveRegistration model to save registrations to scheduled events
 - Add DRF read and create liveRegistration endpoints
 - Allow DRF list action for liveRegistration endpoint
+- Use existing username in jitsi if defined
 
 ### Changed